+++ src/38query.js
@@ -21,6 +21,12 @@ function queryfn(query,oldscope) {
 //		console.log(source, source.data);
 		if(typeof source.data == 'function') {
 			source.getfn = source.data;
+			source.dontcache = source.getfn.dontcache;
+
+//			var prevsource = query.sources[h-1];
+			if(source.joinmode == 'OUTER' || source.joinmode == 'RIGHT' || source.joinmode == 'ANTI') {
+				source.dontcache = false;
+			}
 			source.data = {};
 		}
 //
@@ -36,6 +42,10 @@ function queryfn(query,oldscope) {
 //	if(!oldscope) {
 		preIndex(query);
 //	}
+
+	query.sources.forEach(function(source) {
+			console.log(source.data);
+	});
 
 	// Prepare variables
 	query.data = [];
@@ -119,7 +129,7 @@ preIndex = function(query) {
 		// If there is indexation rule
 //console.log('preIndex', source);
 
-		if(source.optimization == 'ix' && source.onleftfn && source.onrightfn) {
+		if(k > 0 && source.optimization == 'ix' && source.onleftfn && source.onrightfn) {
 			// If there is no table.indices - create it
 			if(query.database.tables[source.tableid]) {
 				if(!query.database.tables[source.tableid].indices) query.database.tables[source.tableid].indices = {};
@@ -139,8 +149,8 @@ preIndex = function(query) {
 				var dataw;
 //				while(source.getfn i<ilen) {
 
-				while((dataw = source.data[i]) || (source.getfn && (dataw = source.getfn(i))) || i<ilen) {
-					if(source.getfn && !source.getfn.dontcache) source.data[i] = dataw;
+				while((dataw = source.data[i]) || (source.getfn && (dataw = source.getfn(i))) || (i<ilen)) {
+					if(source.getfn && !source.dontcache) source.data[i] = dataw;
 //					scope[tableid] = dataw;
 
 //				for(var i=0, ilen=source.data.length; i<ilen; i++) {
@@ -179,7 +189,14 @@ preIndex = function(query) {
 				// Prepare scope
 				var scope = {};
 				// Walking on each source line
-				for(var i=0, ilen=source.data.length; i<ilen; i++) {
+				var i = 0;
+				var ilen = source.data.length;
+				var dataw;
+//				while(source.getfn i<ilen) {
+
+				while((dataw = source.data[i]) || (source.getfn && (dataw = source.getfn(i))) || (i<ilen)) {
+					if(source.getfn && !source.dontcache) source.data[i] = dataw;
+//				for(var i=0, ilen=source.data.length; i<ilen; i++) {
 					scope[source.alias || source.tableid] = source.data[i];
 					// Create index entry
 					var addr = source.wxleftfn(scope, query.params, alasql);
@@ -188,6 +205,7 @@ preIndex = function(query) {
 						group = source.ix [addr] = []; 
 					}
 					group.push(source.data[i]);
+					i++;
 				}
 //					query.database.tables[source.tableid].indices[hash(source.wxleftfns+'`'+source.onwherefns)] = source.ix;
 				query.database.tables[source.tableid].indices[hash(source.wxleftfns+'`')] = source.ix;
@@ -206,13 +224,31 @@ preIndex = function(query) {
 			}		
 
 		// If there is no any optimization than apply srcwhere filter
-		} else if(source.srcwherefns) {
+		} else if(source.srcwherefns && !source.dontcache) {
 			if(source.data) {
 				var scope = {};
+				// TODO!!!!! Data as Function
+
 				source.data = source.data.filter(function(r) {
 					scope[source.alias] = r;
 					return source.srcwherefn(scope, query.params, alasql);
 				});
+
+				var scope = {};
+				var i = 0;
+				var ilen = source.data.length;
+				var dataw;
+				var res = [];
+//				while(source.getfn i<ilen) {
+
+				while((dataw = source.data[i]) || (source.getfn && (dataw = source.getfn(i))) || (i<ilen)) {
+					if(source.getfn && !source.dontcache) source.data[i] = dataw;
+					scope[source.alias] = r;
+					if(source.srcwherefn(scope, query.params, alasql)) res.push(dataw);
+					i++;
+				}
+				source.data = res;
+
 			} else {
 				source.data = [];
 			};